+++ internal/certificatetransparency/ct-watcher.go
@@ -286,7 +286,7 @@ func (w *Watcher) CreateIndexFile(filePath string) error {
 				continue
 			}
 
-			metrics.index[transparencyLog.URL] = int64(sth.TreeSize)
+			metrics.index[transparencyLog.URL] = sth.TreeSize
 		}
 	}
 	w.cancelFunc()
@@ -304,7 +304,7 @@ type worker struct {
 	operatorName string
 	ctURL        string
 	entryChan    chan models.Entry
-	ctIndex      int64
+	ctIndex      uint64
 	mu           sync.Mutex
 	running      bool
 	cancel       context.CancelFunc
@@ -394,14 +394,14 @@ func (w *worker) runWorker(ctx context.Context) error {
 			return errFetchingSTHFailed
 		}
 		// Start at the latest STH to skip all the past certificates
-		w.ctIndex = int64(sth.TreeSize)
+		w.ctIndex = sth.TreeSize
 	}
 
 	certScanner := scanner.NewScanner(jsonClient, scanner.ScannerOptions{
 		FetcherOptions: scanner.FetcherOptions{
 			BatchSize:     100,
 			ParallelFetch: 1,
-			StartIndex:    w.ctIndex,
+			StartIndex:    int64(w.ctIndex),
 			Continuous:    true,
 		},
 		Matcher:     scanner.MatchAll{},
